+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.4
+
+- feat(pkg): Refresh stale cached rules when rules file changes (#5)
+  - Restarting the analysis server is no longer required after modifying the rules file.
+
 ## 0.0.3
 
 - Fix broken image links in README.md.